requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dateutil>=2.9.0
